+++ .clang-format
@@ -56,7 +56,7 @@ IncludeBlocks: Regroup
 IncludeCategories:
   - Regex:           '(<|")((C|c)atch.*)/'
     Priority:        5
-  - Regex:           '(<|")((Archive|Data|Specification|StorageDevice|TrustedPeripheral|Error|MockDevice).*)/'
+  - Regex:           '(<|")((Archive|Messaging|Specification|StorageDevice|TrustedPeripheral|Error|MockDevice).*)/'
     Priority:        2
   - Regex:           '<([[:alnum:].]|_|/)+\..*>'
     Priority:        3
