+++ src/lib/Messaging/CMakeLists.txt
@@ -1,8 +1,8 @@
-add_library(Data)
+add_library(Messaging)
 
-target_compile_features(Data PRIVATE cxx_std_20)
+target_compile_features(Messaging PRIVATE cxx_std_20)
 
-target_sources(Data
+target_sources(Messaging
     PRIVATE
         ComPacket.hpp
         Value.cpp
@@ -19,5 +19,5 @@ target_sources(Data
         Method.hpp
 )
 
-target_include_directories(Data INTERFACE "${CMAKE_CURRENT_LIST_DIR}/..")
-target_link_libraries(Data Archive Error)
+target_include_directories(Messaging INTERFACE "${CMAKE_CURRENT_LIST_DIR}/..")
+target_link_libraries(Messaging Archive Error)
