+++ src/lib/Messaging/Token.cpp
