+++ src/lib/MockDevice/MockSession.hpp
@@ -1,6 +1,6 @@
 #include "MockSecurityProvider.hpp"
 
-#include <Data/NativeTypes.hpp>
+#include <Messaging/NativeTypes.hpp>
 #include <TrustedPeripheral/Method.hpp>
 
 #include <cstdint>
