+++ src/lib/MockDevice/MockSetupLayer.cpp
@@ -1,7 +1,7 @@
 #include "MockSetupLayer.hpp"
 
 #include <Archive/Conversion.hpp>
-#include <Data/SetupPackets.hpp>
+#include <Messaging/SetupPackets.hpp>
 
 namespace sedmgr {
