+++ src/lib/Specification/Module.hpp
@@ -2,8 +2,8 @@
 
 #include "Common/TableDesc.hpp"
 
-#include <Data/NativeTypes.hpp>
-#include <Data/Type.hpp>
+#include <Messaging/NativeTypes.hpp>
+#include <Messaging/Type.hpp>
 
 #include <memory>
 #include <optional>
