+++ src/lib/TrustedPeripheral/Method.hpp
@@ -3,8 +3,8 @@
 #include <async++/task.hpp>
 
 #include <Archive/Types/ValueToNative.hpp>
-#include <Data/Method.hpp>
-#include <Data/NativeTypes.hpp>
+#include <Messaging/Method.hpp>
+#include <Messaging/NativeTypes.hpp>
 
 #include <algorithm>
