+++ src/lib/TrustedPeripheral/Session.cpp
@@ -4,8 +4,8 @@
 #include <async++/join.hpp>
 
 #include <Archive/Conversion.hpp>
-#include <Data/ComPacket.hpp>
-#include <Data/TokenStream.hpp>
+#include <Messaging/ComPacket.hpp>
+#include <Messaging/TokenStream.hpp>
 #include <Error/Exception.hpp>
 #include <Specification/Core/CoreModule.hpp>
 #include <Specification/Opal/OpalModule.hpp>
