+++ src/lib/TrustedPeripheral/TrustedPeripheral.hpp
@@ -4,7 +4,7 @@
 #include "ModuleCollection.hpp"
 #include <async++/task.hpp>
 
-#include <Data/SetupPackets.hpp>
+#include <Messaging/SetupPackets.hpp>
 #include <StorageDevice/NvmeDevice.hpp>
 
 #include <chrono>
