+++ test/Messaging/TestNativeTypes.cpp
@@ -1,4 +1,4 @@
-#include <Data/NativeTypes.hpp>
+#include <Messaging/NativeTypes.hpp>
 
 #include <catch2/catch_test_macros.hpp>
