+++ test/Messaging/TestValue.cpp
@@ -1,5 +1,5 @@
 #include <Archive/Conversion.hpp>
-#include <Data/Value.hpp>
+#include <Messaging/Value.hpp>
 
 #include <cereal/archives/json.hpp>
